% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polls.R
\name{read_polls}
\alias{read_polls}
\title{read_polls}
\usage{
read_polls(tipo = "", ccaa = "", prov = "", mun = "", pollster = "")
}
\arguments{
\item{tipo}{El tipo de elección que se quiere descargar. Pueden ser de tres tipos: "generales", "autonómicas" o "municipales".}

\item{ccaa}{La Comunidad Autónoma donde se realiza la encuesta.}

\item{prov}{La Provincia donde se realiza la encuesta.}

\item{mun}{El municipio donde se realiza la encuesta.}

\item{pollster}{Encuestadora y medio de comunicación que publica la encuesta.}
}
\value{
Dataframe con los datos de las encuestas especificadas.
}
\description{
Esta función lee las encuestas recopiladas por TheElectoralReport en función de los filtros escogidos y los guarda en un directorio especificado.
}
